script 6-size

+++ hello_world/6-size.c
@@ -2,7 +2,8 @@
 
 /**
  * main - Entry point
- * return : Always 0 (sucess)
+ *
+ * Return : Always 0 (sucess)
  */
 
 int main(void)
@@ -13,10 +14,10 @@ int main(void)
 		long long int d;
 		char e;
 
-		printf("Size of a char: %lu byte(s)\n", (unsigned int)sizeof(e));
-		printf("Size of an int: %lu byte(s)\n", (unsigned int)sizeof(a));
-		printf("Size of a long int: %lu byte(s)\n", (unsigned int)sizeof(c));
-		printf("Size of a long long int: %lu byte(s)\n", (unsigned int)sizeof(d));
-		printf("Size of a float: %lu byte(s)\n", (unsigned int)sizeof(b));
+		printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(e));
+		printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(a));
+		printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(c));
+		printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(d));
+		printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(b));
 		return (0);
 }
